Backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-generativeai==0.3.2

# Database
pymongo==4.6.1
motor==3.3.2

# Video processing
opencv-python==4.9.0.80
scenedetect[opencv]==0.6.3
Pillow==10.2.0

# Audio processing
pydub==0.25.1

# Task queue (for async processing)
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1

# Progress tracking
tqdm==4.66.1
